Add isPunished and getBlockBytes utility functions
